+++ Finjector.Web/ClientApp/src/App.tsx
@@ -6,6 +6,7 @@ import logo from "./logo.svg";
 
 import { useUserInfoQuery } from "./queries/userQueries";
 import Landing from "./pages/Landing";
+import About from "./pages/About";
 import Selected from "./pages/Selected";
 import Entry from "./pages/Entry";
 import Paste from "./pages/Paste";
@@ -36,34 +37,6 @@ function App() {
         </Routes>
       </div>
     </BrowserRouter>
-  );
-}
-
-function About() {
-  const user = useUserInfoQuery();
-
-  return (
-    <div className="App">
-      <header className="App-header">
-        <p>
-          {user.isSuccess &&
-            user.data[
-              "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/name"
-            ]}
-        </p>
-        <p>
-          This application was developed by the CAES Dean's Offic Computer
-          Resource Unit! <hr />
-          It uses the Aggie Enterprise Graph QL to build a COA from the campus
-          data source
-        </p>
-        <Link className="App-link" to="/">
-          Back home
-        </Link>
-        <hr />
-        <a href="https://caeshelp.ucdavis.edu/?appname=Finjector">Help</a>
-      </header>
-    </div>
   );
 }
